Ensure webui "channel" is always of corresponding enum
The "channel" was just the string from the config file, which will never equal the enum unless via case-insensitive comparison
