+++ server/src/main/kotlin/suwayomi/tachidesk/graphql/types/WebUIUpdateType.kt
@@ -1,18 +1,20 @@
 package suwayomi.tachidesk.graphql.types
 
+import suwayomi.tachidesk.server.serverConfig
+
 data class AboutWebUI(
-    val channel: String,
+    val channel: WebUIChannel,
     val tag: String,
 )
 
 data class WebUIUpdateCheck(
-    val channel: String,
+    val channel: WebUIChannel,
     val tag: String,
     val updateAvailable: Boolean,
 )
 
 data class WebUIUpdateInfo(
-    val channel: String,
+    val channel: WebUIChannel,
     val tag: String,
 )
 
@@ -28,3 +30,68 @@ data class WebUIUpdateStatus(
     val state: UpdateState,
     val progress: Int,
 )
+
+enum class WebUIInterface {
+    BROWSER,
+    ELECTRON,
+    ;
+
+    companion object {
+        fun from(value: String): WebUIInterface = entries.find { it.name.lowercase() == value.lowercase() } ?: BROWSER
+    }
+}
+
+enum class WebUIChannel {
+    BUNDLED, // the default webUI version bundled with the server release
+    STABLE,
+    PREVIEW,
+    ;
+
+    companion object {
+        fun from(channel: String): WebUIChannel = entries.find { it.name.lowercase() == channel.lowercase() } ?: STABLE
+
+        fun doesConfigChannelEqual(channel: WebUIChannel): Boolean = serverConfig.webUIChannel.value.equals(channel.name, true)
+    }
+}
+
+enum class WebUIFlavor(
+    val uiName: String,
+    val repoUrl: String,
+    val versionMappingUrl: String,
+    val latestReleaseInfoUrl: String,
+    val baseFileName: String,
+) {
+    WEBUI(
+        "WebUI",
+        "https://github.com/Suwayomi/Suwayomi-WebUI-preview",
+        "https://raw.githubusercontent.com/Suwayomi/Suwayomi-WebUI/master/versionToServerVersionMapping.json",
+        "https://api.github.com/repos/Suwayomi/Suwayomi-WebUI-preview/releases/latest",
+        "Suwayomi-WebUI",
+    ),
+
+    VUI(
+        "VUI",
+        "https://github.com/Suwayomi/Suwayomi-VUI",
+        "https://raw.githubusercontent.com/Suwayomi/Suwayomi-VUI/main/versionToServerVersionMapping.json",
+        "https://api.github.com/repos/Suwayomi/Suwayomi-VUI/releases/latest",
+        "Suwayomi-VUI-Web",
+    ),
+
+    CUSTOM(
+        "Custom",
+        "repoURL",
+        "versionMappingUrl",
+        "latestReleaseInfoURL",
+        "baseFileName",
+    ),
+    ;
+
+    companion object {
+        val default: WebUIFlavor = WEBUI
+
+        fun from(value: String): WebUIFlavor = entries.find { it.uiName == value } ?: default
+
+        val current: WebUIFlavor
+            get() = from(serverConfig.webUIFlavor.value)
+    }
+}
